curso-4-9-fiap-ml-modelos-em-apis/pyproject.toml
[project]
name = "curso-4-9-fiap-ml-modelos-em-apis"
version = "0.1.0"
description = ""
authors = [
    {name = "Sóstenes Souza",email = "sostenes198br@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "joblib (>=1.5.2,<2.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
